docs: update README and all documentation to reflect current implementation
- README.md:
  - Update frontend stack entry: recharts → lightweight-charts v5
  - Update rate limit defaults: RATE_LIMIT_PER_SECOND 60→300, BURST 30→200
  - Update Quick Start: use SQLite DATABASE_URL; auto-migration note replaces
    manual migrate step; fix frontend .env.example path
  - Update Project Structure: remove deleted files, add new components/hooks
  - Fix duplicate sync:cre-abi entry in Commands table
  - Add "Recent Improvements" section documenting UI overhaul, codebase cleanup,
    and backend stability changes
  - Add new Troubleshooting entries: port-in-use, SQLite DB lock, rate limiting,
    wrong network prompt
- docs/frontend-project.md:
  - Update tech stack: recharts → lightweight-charts v5; Framer Motion → CSS
  - Add charting note explaining removal of ohlcv-chart.tsx
  - Update .env.example path to config/frontend.env.example
  - Update component file structure to match current implementation including all
    new components, hooks, and removed dead files
  - Add network guard description with window.ethereum and chainChanged details
  - Add TVChart section in market detail specification
  - Add ETH formatting convention (formatEth from @/lib/utils)
- docs/smart-contracts-and-database.md:
  - Remove PostgreSQL reference from document description
  - Add SQLite/PostgreSQL note in database schema section explaining zero-config
    local dev and production PostgreSQL compatibility
  - Document migration stub files (002–006)
- docs/phpe-and-hybrid-prediction.md:
  - Add HybridPredictor constructor note (no cache parameter after cleanup)
  - Clarify distinction between PHPE context cache (phpe_ctx RwLock) and
    application-level prediction result cache (Cache with TTL)
  - Add implementation note about PredictionService caching strategy

Made-with: Cursor

+++ backend-rust/src/models.rs
@@ -244,7 +244,7 @@ pub struct MarketResolution {
     pub id: i64,
     pub market_id: i64,
     pub resolution_type: String,
-    pub outcome: i16,
+    pub outcome: i64,
     pub confidence: Option<f64>,
     pub source: Option<String>,
     pub raw_value: Option<f64>,
